python/requirements.txt
-i https://pypi.org/simple
paho-mqtt==2.1.0; python_version >= '3.7'
